pyproject.toml
[project]
name = "alpha-zero-chess"
version = "0.1.0"
description = "AlphaZero chess engine implementation in PyTorch"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autograd>=1.8.0",
    "jupytext>=1.17.3",
    "matplotlib>=3.10.6",
    "nbconvert>=7.16.6",
    "notebook>=7.4.5",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "python-chess>=1.2.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
